+++ include/nuttx/can.h
@@ -33,6 +33,8 @@
 #include <sys/socket.h>
 #include <stdint.h>
 
+#include <nuttx/can/can_common.h>
+
 /****************************************************************************
  * Pre-processor Definitions
  ****************************************************************************/
